# web/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comic Reader API",
    "description": "Self-hostable comic reader. See comic-reader-spec.md.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/libraries/{slug}/removed": {
      "get": {
        "tags": [
          "api::reconcile"
        ],
        "operationId": "list_removed",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemovedListView"
                }
              }
            }
          },
          "403": {
            "description": "admin only"
          },
          "404": {
            "description": "library not found"
          }
        }
      }
    },
...
    "/series/{series_slug}/issues/{issue_slug}/confirm-removal": {
      "post": {
        "tags": [
          "api::reconcile"
        ],
        "operationId": "confirm_issue",
        "parameters": [
          {
            "name": "series_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "issue_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "confirmed"
          },
          "403": {
            "description": "admin only"
          },
          "404": {
            "description": "issue not found"
          }
        }
      }
    },
...
    "/series/{series_slug}/issues/{issue_slug}/restore": {
      "post": {
        "tags": [
          "api::reconcile"
        ],
        "operationId": "restore_issue",
        "parameters": [
          {
            "name": "series_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "issue_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "restored"
          },
          "403": {
            "description": "admin only"
          },
          "404": {
            "description": "issue not found"
          },
          "409": {
            "description": "file is still missing on disk"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RemovedIssueView": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "series_id",
          "series_slug",
          "file_path",
          "removed_at"
        ],
        "properties": {
          "file_path": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "removal_confirmed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "removed_at": {
            "type": "string"
          },
          "series_id": {
            "type": "string"
          },
          "series_slug": {
            "type": "string",
            "description": "URL slug of the parent series — paired with `slug` it forms the\n`/series/{series_slug}/issues/{slug}/(restore|confirm-removal)` path."
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "RemovedListView": {
        "type": "object",
        "required": [
          "issues",
          "series"
        ],
        "properties": {
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RemovedIssueView"
            }
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RemovedSeriesView"
            }
          }
        }
      },
      "RemovedSeriesView": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "removed_at"
        ],
        "properties": {
          "folder_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "removal_confirmed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "removed_at": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
    }
  }
}